feat : 검색 조건 초기 filter Factory 생성
- DTO 내부 반환 로직에서 Factory로 분리
